pc: initialize legacy hotplug only for 2.6 and older machine types
on old machine types CPU hotplug was uncondtionally
enabled since it was introduced, consuming IO ports
and providing AML regardles of whether it was actually
in use or not. Keep it so for 2.6 and older machines.

New machine types will have an option to turn CPU
hotplug on if it's needed while by default it stays
disabled not consuming extra RAM/IO resources.

Signed-off-by: Igor Mammedov <imammedo@redhat.com>