apic: add deliver_msi to APICCommonClass
The only way to send interrupts from outside of APIC devices is to use
the MSI-inspired memory mapped interface.  The interface is not
sufficient because interrupt remapping in extended mode can exceed 8 bit
destination ids.

IR unit stores excessive bits in the high word of MSI address register,
but using a memory mapped interface would be wasteful.

The proper way to deliver interrupts would be to design a bus (QPI), but
that is a big undertaking when we can just add one callback to the APIC
class that delivers extended MSIs.

Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>