apic: add deliver_msi to APICCommonClass
The only way to send interrupts from outside of APIC devices is to use
the MSI-inspired memory mapped interface.  The interface is not
sufficient because interrupt remapping in extended mode can exceed 8 bit
destination ids.

The proper way to deliver interrupts would be to design a bus (QPI), but
that is a big undertaking.

Real IR unit stores excessive bits in the hight work of MSI address
register, which would be bit [63:40] in memory address space, so I was
considering two options:
 - a new special address space for APIC to handle MSI-compatible writes
 - one callback to the APIC class that delivers extended MSIs

This patch is the second one, because it seemed easier.

Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>